+++ src/main/java/de/tobi1craft/concacity/client/ConcacityClient.java
@@ -4,7 +4,7 @@ import de.tobi1craft.concacity.client.gui.ModGUIs;
 import de.tobi1craft.concacity.client.key.ModKeys;
 import de.tobi1craft.concacity.client.util.ClientChannels;
 import de.tobi1craft.concacity.entity.ModEntities;
-import de.tobi1craft.concacity.client.entity.HelperRenderer;
+import de.tobi1craft.concacity.client.entity.HelperMinerRenderer;
 import net.fabricmc.api.ClientModInitializer;
 import net.fabricmc.api.EnvType;
 import net.fabricmc.api.Environment;
@@ -21,6 +21,6 @@ public class ConcacityClient implements ClientModInitializer {
         ModGUIs.registerModGUIs();
         ModKeys.registerModKeys();
 
-        EntityRendererRegistry.register(ModEntities.HELPER, HelperRenderer::new);
+        EntityRendererRegistry.register(ModEntities.HELPER, HelperMinerRenderer::new);
     }
 }