+++ src/main/java/de/tobi1craft/concacity/client/entity/HelperMinerModel.java
@@ -1,7 +1,7 @@
 package de.tobi1craft.concacity.client.entity;
 
 import de.tobi1craft.concacity.Concacity;
-import de.tobi1craft.concacity.entity.custom.HelperEntity;
+import de.tobi1craft.concacity.entity.helper.HelperMinerEntity;
 import net.minecraft.util.Identifier;
 import net.minecraft.util.math.MathHelper;
 import software.bernie.geckolib.constant.DataTickets;
@@ -10,24 +10,24 @@ import software.bernie.geckolib.core.animation.AnimationState;
 import software.bernie.geckolib.model.GeoModel;
 import software.bernie.geckolib.model.data.EntityModelData;
 
-public class HelperModel extends GeoModel<HelperEntity> {
+public class HelperMinerModel extends GeoModel<HelperMinerEntity> {
     @Override
-    public Identifier getModelResource(HelperEntity animatable) {
+    public Identifier getModelResource(HelperMinerEntity animatable) {
         return new Identifier(Concacity.ID, "geo/helper.geo.json");
     }
 
     @Override
-    public Identifier getTextureResource(HelperEntity animatable) {
+    public Identifier getTextureResource(HelperMinerEntity animatable) {
         return new Identifier(Concacity.ID, "textures/entity/helper.png");
     }
 
     @Override
-    public Identifier getAnimationResource(HelperEntity animatable) {
+    public Identifier getAnimationResource(HelperMinerEntity animatable) {
         return new Identifier(Concacity.ID, "animations/helper.animation.json");
     }
 
     @Override
-    public void setCustomAnimations(HelperEntity animatable, long instanceId, AnimationState<HelperEntity> animationState) {
+    public void setCustomAnimations(HelperMinerEntity animatable, long instanceId, AnimationState<HelperMinerEntity> animationState) {
         CoreGeoBone head = getAnimationProcessor().getBone("head");
         if (head != null) {
             EntityModelData entityData = animationState.getData(DataTickets.ENTITY_MODEL_DATA);