+++ src/main/java/de/tobi1craft/concacity/client/util/ClientChannels.java
@@ -1,8 +1,6 @@
 package de.tobi1craft.concacity.client.util;
 
 import de.tobi1craft.concacity.Concacity;
-import de.tobi1craft.concacity.client.gui.HelperInventoryGUI;
-import de.tobi1craft.concacity.util.ModPackets;
 import io.wispforest.owo.network.OwoNetChannel;
 
 public class ClientChannels {