+++ src/main/java/de/tobi1craft/concacity/data/ModModelProvider.java
@@ -1,11 +1,13 @@
 package de.tobi1craft.concacity.data;
 
+import de.tobi1craft.concacity.block.ModBlocks;
 import de.tobi1craft.concacity.item.ModItems;
 import net.fabricmc.fabric.api.datagen.v1.FabricDataOutput;
 import net.fabricmc.fabric.api.datagen.v1.provider.FabricModelProvider;
 import net.minecraft.data.client.BlockStateModelGenerator;
 import net.minecraft.data.client.ItemModelGenerator;
 import net.minecraft.data.client.ModelIds;
+import net.minecraft.data.client.Models;
 
 public class ModModelProvider extends FabricModelProvider {
     public ModModelProvider(FabricDataOutput output) {
@@ -14,12 +16,10 @@ public class ModModelProvider extends FabricModelProvider {
 
     @Override
     public void generateBlockStateModels(BlockStateModelGenerator blockStateModelGenerator) {
-        //blockStateModelGenerator.registerCubeAllModelTexturePool(ModBlocks.CONCACITY_BLOCK);
         blockStateModelGenerator.registerParentedItemModel(ModItems.HELPER_SPAWN_EGG, ModelIds.getMinecraftNamespacedItem("template_spawn_egg"));
     }
 
     @Override
     public void generateItemModels(ItemModelGenerator itemModelGenerator) {
-        //itemModelGenerator.register(ModItems.CONCACITY, Models.GENERATED);
     }
 }