+++ src/main/java/de/tobi1craft/concacity/entity/ModEntities.java
@@ -1,7 +1,7 @@
 package de.tobi1craft.concacity.entity;
 
 import de.tobi1craft.concacity.Concacity;
-import de.tobi1craft.concacity.entity.custom.HelperEntity;
+import de.tobi1craft.concacity.entity.helper.HelperMinerEntity;
 import net.fabricmc.fabric.api.object.builder.v1.entity.FabricDefaultAttributeRegistry;
 import net.fabricmc.fabric.api.object.builder.v1.entity.FabricEntityTypeBuilder;
 import net.fabricmc.fabric.api.object.builder.v1.world.poi.PointOfInterestHelper;
@@ -20,7 +20,7 @@ import java.util.Optional;
 
 public class ModEntities {
 
-    public static final EntityType<HelperEntity> HELPER = Registry.register(Registries.ENTITY_TYPE, new Identifier(Concacity.ID, "helper"), FabricEntityTypeBuilder.create(SpawnGroup.MISC, HelperEntity::new)
+    public static final EntityType<HelperMinerEntity> HELPER = Registry.register(Registries.ENTITY_TYPE, new Identifier(Concacity.ID, "helper"), FabricEntityTypeBuilder.create(SpawnGroup.MISC, HelperMinerEntity::new)
             .dimensions(EntityDimensions.fixed(1.5f, 1.75f)).build());
     @SuppressWarnings("unused")
     public static final PointOfInterestType TREE = PointOfInterestHelper.register(new Identifier(Concacity.ID, "tree"), 32, 100, Blocks.OAK_LOG);
@@ -29,6 +29,6 @@ public class ModEntities {
 
     public static void registerModEntities() {
         Concacity.LOGGER.info("registering mod entities");
-        FabricDefaultAttributeRegistry.register(ModEntities.HELPER, HelperEntity.setAttributes());
+        FabricDefaultAttributeRegistry.register(ModEntities.HELPER, HelperMinerEntity.setAttributes());
     }
 }